Remove stuff to do with multiple selection, copy & paste.
