+++ trnsysGUI/PortItemBase.py
@@ -111,10 +111,8 @@ class PortItemBase(QGraphicsEllipseItem):
                     e = conn.segments[-1]
                     e.setLine(e.line().p1().x(), e.line().p1().y(), self.scenePos().x(), self.scenePos().y())
 
-                # global pasting
-                if not self.parent.parent.parent().pasting:
-                    for s in conn.segments:
-                        s.updateGrad()
+                for s in conn.segments:
+                    s.updateGrad()
 
         if change == self.ItemScenePositionHasChanged and self.parent.parent.parent().editorMode == 1:
             for conn in self.connectionList:
